Fix balance hours (to decimal instead of int)

+++ src/Basic.Model/Balance.cs
@@ -32,11 +32,11 @@ namespace Basic.Model
         /// Gets or sets the defined standard allowance for this year, in hours.
         /// </summary>
         [Required]
-        public int Allowed { get; set; }
+        public decimal Allowed { get; set; }
 
         /// <summary>
         /// Gets or sets the transfered amount to add to the balance, in hours.
         /// </summary>
-        public int Transfered { get; set; }
+        public decimal Transfered { get; set; }
     }
 }
